backend/requirements.txt
fastapi
uvicorn
sentence-transformers
torch
mysql-connector-python
nltk
